requirements.txt
asgiref==3.2.10
boto3==1.14.38
botocore==1.17.38
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
Django==3.1
django-allauth==0.42.0
django-cors-headers==3.4.0
django-filter==2.3.0
django-rest-swagger==2.2.0
django-storages==1.9.1
djangorestframework==3.11.1
docutils==0.15.2
idna==2.10
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
MarkupSafe==1.1.1
mysqlclient==2.0.1
oauthlib==3.1.0
openapi-codec==1.3.2
Pillow==7.2.0
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.10